Moved load code to a file.
